Move a monitor_indexes::Db::get_index_target_type into a db index actor
This is a part of #3.

The change moves get_index_target_type method from monitor_indexes into
a db actor. This will also remove monitor_indexes empty Db struct. Next
paches will move other methods into a db actor.
